Fix repository ids in example settings.xml
This avoids a nasty gotcha in maven's repository handling. The repository
ids in the distributionManagement section of the pom don't need to be
unique. This allows us to use the same <server> configuration in the
settings to deploy to both snapshot and release repositories.

However, the repository ids in the settings.xml (i.e. the ones changed
in this patch) *do* need to be unique. If they are not unique, then maven
will try to resolve everything from repo1.maven.org.